[SPIRV-INTERMEDIATE] Using SPIRV as output of Compile API of CommonClang
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/106802
Tested-by: Sotkin, Alexey <alexey.sotkin@intel.com>

+++ opencl-intel/cl_api/frontend_api.h
@@ -80,7 +80,7 @@ struct FESPIRVProgramDescriptor
     // binary container for SPIRV program
     const void*     pSPIRVContainer;
     // the size in bytes of the container pSPIRVContainer
-    unsigned int    uiSPIRVContainerSize;
+    size_t          uiSPIRVContainerSize;
     // A string for compile options
     const char*     pszOptions;
 };
